improved rasterization

+++ src/rendering/point.rs
@@ -2,12 +2,12 @@ use crate::math::vertex::Vertex;
 
 #[derive(Copy, Clone, Debug)]
 pub struct Point{
-    pub x: i32,
-    pub y: i32,
+    pub x: f32,
+    pub y: f32,
 }
 
 impl Point {
-    pub fn new(x: i32, y: i32) -> Point {
+    pub fn new(x: f32, y: f32) -> Point {
         Point { x, y }
     }
 }
